requirements.txt
unsloth
transformers<4.41.0
datasets
trl
accelerate
bitsandbytes
scikit-learn
pandas
pyyaml